// src/shared/enums/messages/success-messages.enum.ts
export enum SuccessMessages {
    CreateSuccess = 'Resource created successfully',
    GetSuccess = 'Resource retrieved successfully',
    UpdateSuccess = 'Resource updated successfully',
    DeleteSuccess = 'Resource deleted successfully',
    GenericSuccess = 'Operation completed successfully',
    UserRemoveSuccess = 'User removed!',
    ProductRemoveSuccess = 'Product removed!'
}